History: the whole DB half moves to database/history_repository
common/history.c was already most of the way there -- file-static prepared
statements, a dt_history_db_* family, and a row callback that decodes each row
into scalars rather than handing a cursor to its caller. This lifts that block
into src/database/history_repository.{h,c} unchanged: all 20 query strings are
byte-identical to master's.

What stayed in common/history.c is the policy, which is all it has left: the
undo group, the preset flag, the mipmap invalidation, the darktable|style|*
tag detach, and the two listing functions -- those keep calling
dt_iop_get_localized_name(), which lives four layers up and must never be
reachable from a layer-1 repository, so the repository hands back
(num, operation, enabled, multi_name) and history.c does the naming.

Three things found on the way, each fixed rather than carried across:

- dt_history_hash_set_mipmap() was declared in history.h, defined nowhere, and
  called from nowhere; its cached statement was finalised in cleanup and never
  prepared. Declaration and statement both dropped.
- dt_history_check_module_exists()'s `enabled` argument was never bound -- the
  query is "WHERE imgid = ?1 AND operation = ?2" and its one caller passed
  FALSE. Gone; the function is dt_history_repository_module_exists(imgid, op).
- the five deletes in dt_history_delete_on_image_ext() ignored every step
  result. As dt_history_repository_delete_all_for_image() they are ANDed into a
  return value, which the caller is still free to ignore.

Removing <sqlite3.h> and caches/image_cache.h from history.h broke three files
that had been reaching dt_pthread_mutex_t and dt_image_cache_get() through it:
control/control.h and views/view.h declare dt_pthread_mutex_t members,
imageio/storage/piwigo.c calls the image cache. All three now name what they
use, per the header rule.

Ratchet: 198 -> 178 handles, 29 -> 28 files.
