+++ src/common/history.c
@@ -39,14 +39,12 @@
     along with darktable.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-#include "database/database.h"
+#include "database/history_repository.h"
 #include "common/thumbnail_notify.h"
 #include "common/history.h"
 #include "system/macros.h"
 #include "system/mem_alloc.h"
 #include "common/logging.h"
-#include "database/sql_debug.h"
-#include "system/dtpthread.h"
 #include "common/history_snapshot.h"
 #include "caches/image_cache.h"
 #include "caches/mipmap_cache.h"
@@ -57,35 +55,6 @@
 #include "widgets/label.h"
 
 #define DT_IOP_ORDER_INFO (dt_get_debug_flags() & DT_DEBUG_IOPORDER)
-
-static sqlite3_stmt *_history_check_module_exists_stmt = NULL;
-static sqlite3_stmt *_history_hash_set_mipmap_stmt = NULL;
-static sqlite3_stmt *_history_get_end_stmt = NULL;
-static sqlite3_stmt *_history_set_end_stmt = NULL;
-static sqlite3_stmt *_history_get_next_num_stmt = NULL;
-
-static sqlite3_stmt *_history_delete_history_stmt = NULL;
-static sqlite3_stmt *_history_delete_masks_stmt = NULL;
-static sqlite3_stmt *_history_shift_history_nums_stmt = NULL;
-static sqlite3_stmt *_history_select_history_stmt = NULL;
-static sqlite3_stmt *_history_select_num_stmt = NULL;
-static sqlite3_stmt *_history_insert_num_stmt = NULL;
-static sqlite3_stmt *_history_update_item_stmt = NULL;
-
-static sqlite3_stmt *_history_auto_presets_stmt = NULL;
-static sqlite3_stmt *_history_auto_presets_legacy_stmt = NULL;
-static sqlite3_stmt *_history_auto_ioporder_stmt = NULL;
-static dt_pthread_mutex_t _history_stmt_mutex;
-static gsize _history_stmt_mutex_inited = 0;
-
-static inline void _history_stmt_mutex_ensure(void)
-{
-  if(g_once_init_enter(&_history_stmt_mutex_inited))
-  {
-    dt_pthread_mutex_init(&_history_stmt_mutex, NULL);
-    g_once_init_leave(&_history_stmt_mutex_inited, 1);
-  }
-}
 
 void dt_history_item_free(gpointer data)
 {
@@ -119,46 +88,7 @@ void dt_history_delete_on_image_ext(int32_t imgid, gboolean undo)
     dt_history_snapshot_undo_create(hist->imgid, &hist->before, &hist->before_history_end);
   }
 
-  sqlite3_stmt *stmt;
-
-  DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                              "DELETE FROM main.history WHERE imgid = ?1",
-                              -1, &stmt, NULL);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  sqlite3_step(stmt);
-  sqlite3_finalize(stmt);
-
-  DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                              "DELETE FROM main.module_order WHERE imgid = ?1",
-                              -1, &stmt, NULL);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  sqlite3_step(stmt);
-  sqlite3_finalize(stmt);
-
-  // clang-format off
-  DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                              "UPDATE main.images"
-                              " SET history_end = 0, aspect_ratio = 0.0"
-                              " WHERE id = ?1",
-                              -1, &stmt, NULL);
-  // clang-format on
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  sqlite3_step(stmt);
-  sqlite3_finalize(stmt);
-
-  DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                              "DELETE FROM main.masks_history WHERE imgid = ?1",
-                              -1, &stmt, NULL);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  sqlite3_step(stmt);
-  sqlite3_finalize(stmt);
-
-  DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                              "DELETE FROM main.history_hash WHERE imgid = ?1",
-                              -1, &stmt, NULL);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  sqlite3_step(stmt);
-  sqlite3_finalize(stmt);
+  dt_history_repository_delete_all_for_image(imgid);
 
   _remove_preset_flag(imgid);
 
@@ -194,598 +124,63 @@ char *dt_history_item_as_string(const char *name, gboolean enabled)
   return g_strconcat(enabled ? "\342\227\217" : "\342\227\213", "  ", name, NULL);
 }
 
+static void _collect_item(void *user_data, const int num, const char *operation,
+                          const gboolean enabled, const char *multi_name)
+{
+  GList **result = (GList **)user_data;
+
+  if(strcmp(operation, "mask_manager") == 0) return;
+
+  char name[512] = { 0 };
+  dt_history_item_t *item = g_malloc(sizeof(dt_history_item_t));
+  item->num = num;
+  item->enabled = enabled;
+
+  if(strcmp(multi_name, "0") == 0)
+    g_snprintf(name, sizeof(name), "%s", dt_iop_get_localized_name(operation));
+  else
+    g_snprintf(name, sizeof(name), "%s %s", dt_iop_get_localized_name(operation), multi_name);
+
+  item->name = g_strdup(name);
+  item->op = g_strdup(operation);
+  *result = g_list_prepend(*result, item);
+}
+
 GList *dt_history_get_items(const int32_t imgid, gboolean enabled)
 {
   GList *result = NULL;
-  sqlite3_stmt *stmt;
-
-  // clang-format off
-  DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                              "SELECT num, operation, enabled, multi_name"
-                              " FROM main.history"
-                              " WHERE imgid=?1"
-                              "   AND num IN (SELECT MAX(num)"
-                              "               FROM main.history hst2"
-                              "               WHERE hst2.imgid=?1"
-                              "                 AND hst2.operation=main.history.operation"
-                              "               GROUP BY multi_priority)"
-                              "   AND enabled in (1, ?2)"
-                              " ORDER BY num DESC",
-                              -1, &stmt, NULL);
-  // clang-format on
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 2, enabled ? 1 : 0);
-
-  while(sqlite3_step(stmt) == SQLITE_ROW)
-  {
-    if(strcmp((const char*)sqlite3_column_text(stmt, 1), "mask_manager") == 0) continue;
-
-    char name[512] = { 0 };
-    dt_history_item_t *item = g_malloc(sizeof(dt_history_item_t));
-    const char *op = (char *)sqlite3_column_text(stmt, 1);
-    item->num = sqlite3_column_int(stmt, 0);
-    item->enabled = sqlite3_column_int(stmt, 2);
-
-    char *mname = g_strdup((gchar *)sqlite3_column_text(stmt, 3));
-
-    if(strcmp(mname, "0") == 0)
-      g_snprintf(name, sizeof(name), "%s", dt_iop_get_localized_name(op));
-    else
-      g_snprintf(name, sizeof(name), "%s %s",
-                 dt_iop_get_localized_name(op),
-                 (char *)sqlite3_column_text(stmt, 3));
-    item->name = g_strdup(name);
-    item->op = g_strdup(op);
-    result = g_list_prepend(result, item);
-
-    dt_free(mname);
-  }
-  sqlite3_finalize(stmt);
+  dt_history_repository_foreach_last_item(imgid, enabled, _collect_item, &result);
   return g_list_reverse(result);   // list was built in reverse order, so un-reverse it
+}
+
+static void _collect_item_string(void *user_data, const int num, const char *operation,
+                                 const gboolean enabled, const char *multi_name)
+{
+  GList **items = (GList **)user_data;
+  char *decorated = NULL;
+
+  if(multi_name && *multi_name && g_strcmp0(multi_name, " ") != 0 && g_strcmp0(multi_name, "0") != 0)
+    decorated = g_strconcat(" ", multi_name, NULL);
+
+  char *iname = dt_history_item_as_string(dt_iop_get_localized_name(operation), enabled);
+  char *name = g_strconcat(iname, decorated ? decorated : "", NULL);
+  *items = g_list_prepend(*items, delete_underscore(name));
+
+  dt_free(iname);
+  dt_free(name);
+  dt_free(decorated);
 }
 
 char *dt_history_get_items_as_string(const int32_t imgid)
 {
   GList *items = NULL;
-  sqlite3_stmt *stmt;
-  // clang-format off
-  DT_DEBUG_SQLITE3_PREPARE_V2(
-      dt_database_get_sqlite3_global(),
-      "SELECT operation, enabled, multi_name"
-      " FROM main.history"
-      " WHERE imgid=?1 ORDER BY num DESC", -1, &stmt, NULL);
-  // clang-format on
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-
-  // collect all the entries in the history from the db
-  while(sqlite3_step(stmt) == SQLITE_ROW)
-  {
-    char *multi_name = NULL;
-    const char *mn = (char *)sqlite3_column_text(stmt, 2);
-
-    if(mn && *mn && g_strcmp0(mn, " ") != 0 && g_strcmp0(mn, "0") != 0)
-      multi_name = g_strconcat(" ", sqlite3_column_text(stmt, 2), NULL);
-
-    char *iname = dt_history_item_as_string
-      (dt_iop_get_localized_name((char *)sqlite3_column_text(stmt, 0)),
-       sqlite3_column_int(stmt, 1));
-
-    char *name = g_strconcat(iname, multi_name ? multi_name : "", NULL);
-    char *clean_name = delete_underscore(name);
-    items = g_list_prepend(items, clean_name);
-
-    dt_free(iname);
-    dt_free(name);
-    dt_free(multi_name);
-  }
-  sqlite3_finalize(stmt);
+  dt_history_repository_foreach_item(imgid, _collect_item_string, &items);
   items = g_list_reverse(items); // list was built in reverse order, so un-reverse it
   char *result = dt_util_glist_to_str("\n", items);
   g_list_free_full(items, dt_free_gpointer);
-  items = NULL;
   return result;
 }
 
-gboolean dt_history_check_module_exists(int32_t imgid, const char *operation, gboolean enabled)
-{
-  gboolean result = FALSE;
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-  if(!_history_check_module_exists_stmt)
-  {
-    // clang-format off
-    DT_DEBUG_SQLITE3_PREPARE_V2(
-      dt_database_get_sqlite3_global(),
-      "SELECT imgid"
-      " FROM main.history"
-      " WHERE imgid= ?1 AND operation = ?2",
-      -1, &_history_check_module_exists_stmt, NULL);
-    // clang-format on
-  }
-  sqlite3_stmt *stmt = _history_check_module_exists_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 2, operation, -1, SQLITE_TRANSIENT);
-  if (sqlite3_step(stmt) == SQLITE_ROW) result = TRUE;
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-
-  return result;
-}
-
-void dt_history_cleanup(void)
-{
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-  if(_history_hash_set_mipmap_stmt)
-  {
-    sqlite3_finalize(_history_hash_set_mipmap_stmt);
-    _history_hash_set_mipmap_stmt = NULL;
-  }
-  if(_history_check_module_exists_stmt)
-  {
-    sqlite3_finalize(_history_check_module_exists_stmt);
-    _history_check_module_exists_stmt = NULL;
-  }
-  if(_history_get_end_stmt)
-  {
-    sqlite3_finalize(_history_get_end_stmt);
-    _history_get_end_stmt = NULL;
-  }
-  if(_history_set_end_stmt)
-  {
-    sqlite3_finalize(_history_set_end_stmt);
-    _history_set_end_stmt = NULL;
-  }
-  if(_history_get_next_num_stmt)
-  {
-    sqlite3_finalize(_history_get_next_num_stmt);
-    _history_get_next_num_stmt = NULL;
-  }
-  if(_history_delete_history_stmt)
-  {
-    sqlite3_finalize(_history_delete_history_stmt);
-    _history_delete_history_stmt = NULL;
-  }
-  if(_history_delete_masks_stmt)
-  {
-    sqlite3_finalize(_history_delete_masks_stmt);
-    _history_delete_masks_stmt = NULL;
-  }
-  if(_history_shift_history_nums_stmt)
-  {
-    sqlite3_finalize(_history_shift_history_nums_stmt);
-    _history_shift_history_nums_stmt = NULL;
-  }
-  if(_history_select_history_stmt)
-  {
-    sqlite3_finalize(_history_select_history_stmt);
-    _history_select_history_stmt = NULL;
-  }
-  if(_history_select_num_stmt)
-  {
-    sqlite3_finalize(_history_select_num_stmt);
-    _history_select_num_stmt = NULL;
-  }
-  if(_history_insert_num_stmt)
-  {
-    sqlite3_finalize(_history_insert_num_stmt);
-    _history_insert_num_stmt = NULL;
-  }
-  if(_history_update_item_stmt)
-  {
-    sqlite3_finalize(_history_update_item_stmt);
-    _history_update_item_stmt = NULL;
-  }
-  if(_history_auto_presets_stmt)
-  {
-    sqlite3_finalize(_history_auto_presets_stmt);
-    _history_auto_presets_stmt = NULL;
-  }
-  if(_history_auto_presets_legacy_stmt)
-  {
-    sqlite3_finalize(_history_auto_presets_legacy_stmt);
-    _history_auto_presets_legacy_stmt = NULL;
-  }
-  if(_history_auto_ioporder_stmt)
-  {
-    sqlite3_finalize(_history_auto_ioporder_stmt);
-    _history_auto_ioporder_stmt = NULL;
-  }
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-}
-
-int32_t dt_history_get_end(const int32_t imgid)
-{
-  if(imgid <= 0) return 0;
-
-  int32_t end = 0;
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-  if(!_history_get_end_stmt)
-  {
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                "SELECT history_end FROM main.images WHERE id=?1", -1,
-                                &_history_get_end_stmt, NULL);
-  }
-  sqlite3_stmt *stmt = _history_get_end_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  if(sqlite3_step(stmt) == SQLITE_ROW && sqlite3_column_type(stmt, 0) != SQLITE_NULL)
-    end = sqlite3_column_int(stmt, 0);
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-
-  return end;
-}
-
-gboolean dt_history_set_end(const int32_t imgid, const int32_t history_end)
-{
-  if(imgid <= 0) return FALSE;
-
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-  if(!_history_set_end_stmt)
-  {
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                "UPDATE main.images SET history_end = ?1 WHERE id = ?2", -1,
-                                &_history_set_end_stmt, NULL);
-  }
-  sqlite3_stmt *stmt = _history_set_end_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, history_end);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 2, imgid);
-  const gboolean ok = (sqlite3_step(stmt) == SQLITE_DONE);
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-  return ok;
-}
-
-int32_t dt_history_db_get_next_history_num(const int32_t imgid)
-{
-  if(imgid <= 0) return 0;
-
-  int32_t next_num = 0;
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-  if(!_history_get_next_num_stmt)
-  {
-    // clang-format off
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                "SELECT IFNULL(MAX(num)+1, 0) FROM main.history"
-                                " WHERE imgid = ?1",
-                                -1, &_history_get_next_num_stmt, NULL);
-    // clang-format on
-  }
-  sqlite3_stmt *stmt = _history_get_next_num_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  if(sqlite3_step(stmt) == SQLITE_ROW)
-    next_num = sqlite3_column_int(stmt, 0);
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-  return next_num;
-}
-
-gboolean dt_history_db_delete_history(const int32_t imgid)
-{
-  if(imgid <= 0) return FALSE;
-
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-  if(!_history_delete_history_stmt)
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                "DELETE FROM main.history WHERE imgid = ?1", -1,
-                                &_history_delete_history_stmt, NULL);
-  sqlite3_stmt *stmt = _history_delete_history_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  const gboolean ok = (sqlite3_step(stmt) == SQLITE_DONE);
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-  return ok;
-}
-
-gboolean dt_history_db_delete_masks_history(const int32_t imgid)
-{
-  if(imgid <= 0) return FALSE;
-
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-  if(!_history_delete_masks_stmt)
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                "DELETE FROM main.masks_history WHERE imgid = ?1", -1,
-                                &_history_delete_masks_stmt, NULL);
-  sqlite3_stmt *stmt = _history_delete_masks_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  const gboolean ok = (sqlite3_step(stmt) == SQLITE_DONE);
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-  return ok;
-}
-
-gboolean dt_history_db_shift_history_nums(const int32_t imgid, const int delta)
-{
-  if(imgid <= 0 || delta == 0) return TRUE;
-
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-  if(!_history_shift_history_nums_stmt)
-  {
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                "UPDATE main.history SET num = num + ?2 WHERE imgid = ?1", -1,
-                                &_history_shift_history_nums_stmt, NULL);
-  }
-  sqlite3_stmt *stmt = _history_shift_history_nums_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 2, delta);
-  const gboolean ok = (sqlite3_step(stmt) == SQLITE_DONE);
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-  return ok;
-}
-
-gboolean dt_history_db_delete_dev_history(const int32_t imgid)
-{
-  if(imgid <= 0) return FALSE;
-  gboolean ok = TRUE;
-  ok &= dt_history_db_delete_history(imgid);
-  ok &= dt_history_db_delete_masks_history(imgid);
-  return ok;
-}
-
-gboolean dt_history_db_write_history_item(const int32_t imgid, const int num, const char *operation, const void *op_params,
-                                         const int op_params_size, const int module_version, const gboolean enabled,
-                                         const void *blendop_params, const int blendop_params_size,
-                                         const int blendop_version, const int multi_priority, const char *multi_name)
-{
-  if(imgid <= 0 || num < 0 || IS_NULL_PTR(operation)) return FALSE;
-
-  gboolean ok = TRUE;
-
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-
-  if(!_history_select_num_stmt)
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                "SELECT num FROM main.history WHERE imgid = ?1 AND num = ?2", -1,
-                                &_history_select_num_stmt, NULL);
-  sqlite3_stmt *stmt = _history_select_num_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 2, num);
-  if(sqlite3_step(stmt) != SQLITE_ROW)
-  {
-    if(!_history_insert_num_stmt)
-      DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                  "INSERT INTO main.history (imgid, num) VALUES (?1, ?2)", -1,
-                                  &_history_insert_num_stmt, NULL);
-    stmt = _history_insert_num_stmt;
-    sqlite3_reset(stmt);
-    sqlite3_clear_bindings(stmt);
-    DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-    DT_DEBUG_SQLITE3_BIND_INT(stmt, 2, num);
-    ok &= (sqlite3_step(stmt) == SQLITE_DONE);
-  }
-
-  if(!_history_update_item_stmt)
-    // clang-format off
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                "UPDATE main.history"
-                                " SET operation = ?1, op_params = ?2, module = ?3, enabled = ?4, "
-                                "     blendop_params = ?7, blendop_version = ?8, multi_priority = ?9, multi_name = ?10"
-                                " WHERE imgid = ?5 AND num = ?6",
-                                -1, &_history_update_item_stmt, NULL);
-    // clang-format on
-  stmt = _history_update_item_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 1, operation, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_BLOB(stmt, 2, op_params, op_params_size, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 3, module_version);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 4, enabled);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 5, imgid);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 6, num);
-  DT_DEBUG_SQLITE3_BIND_BLOB(stmt, 7, blendop_params, blendop_params_size, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 8, blendop_version);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 9, multi_priority);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 10, multi_name ? multi_name : "", -1, SQLITE_TRANSIENT);
-  ok &= (sqlite3_step(stmt) == SQLITE_DONE);
-
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-  return ok;
-}
-
-void dt_history_db_foreach_history_row(const int32_t imgid, dt_history_db_row_cb cb, void *user_data)
-{
-  if(imgid <= 0 || IS_NULL_PTR(cb)) return;
-
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-  if(!_history_select_history_stmt)
-  {
-    // clang-format off
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                "SELECT imgid, num, module, operation,"
-                                "       op_params, enabled, blendop_params,"
-                                "       blendop_version, multi_priority, multi_name"
-                                " FROM main.history"
-                                " WHERE imgid = ?1"
-                                " ORDER BY num",
-                                -1, &_history_select_history_stmt, NULL);
-    // clang-format on
-  }
-
-  sqlite3_stmt *stmt = _history_select_history_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-
-  while(sqlite3_step(stmt) == SQLITE_ROW)
-  {
-    const int32_t id = sqlite3_column_int(stmt, 0);
-    const int num = sqlite3_column_int(stmt, 1);
-    const int modversion = sqlite3_column_int(stmt, 2);
-    const char *operation = (const char *)sqlite3_column_text(stmt, 3);
-    const void *module_params = sqlite3_column_blob(stmt, 4);
-    const gboolean enabled = sqlite3_column_int(stmt, 5) != 0; // ensure casting to gboolean
-    const void *blendop_params = sqlite3_column_blob(stmt, 6);
-    const int blendop_version = sqlite3_column_int(stmt, 7);
-    const int multi_priority = sqlite3_column_int(stmt, 8);
-    const char *multi_name = (const char *)sqlite3_column_text(stmt, 9);
-    const int param_length = sqlite3_column_bytes(stmt, 4);
-    const int bl_length = sqlite3_column_bytes(stmt, 6);
-
-    cb(user_data, id, num, modversion, operation, module_params, param_length, enabled,
-       blendop_params, bl_length, blendop_version, multi_priority, multi_name, "");
-  }
-
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-}
-
-void dt_history_db_foreach_auto_preset_row(const int32_t imgid, const dt_image_t *image, const char *workflow_preset,
-                                          const int iformat, const int excluded, dt_history_db_row_cb cb, void *user_data)
-{
-  if(imgid <= 0 || IS_NULL_PTR(image) || !workflow_preset || IS_NULL_PTR(cb)) return;
-
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-
-  const gboolean use_modern_presets = (image->flags & DT_IMAGE_NO_LEGACY_PRESETS);
-  sqlite3_stmt **stmt_ptr = use_modern_presets ? &_history_auto_presets_stmt : &_history_auto_presets_legacy_stmt;
-  if(!*stmt_ptr)
-  {
-    const char *table = use_modern_presets ? "data.presets" : "main.legacy_presets";
-
-    // clang-format off
-    char *query = g_strdup_printf(
-      " SELECT ?1, 0, op_version, operation, op_params,"
-      "       enabled, blendop_params, blendop_version, multi_priority, multi_name, name"
-      " FROM %s"
-      " WHERE ( (autoapply=1"
-      "          AND ((?2 LIKE model AND ?3 LIKE maker) OR (?4 LIKE model AND ?5 LIKE maker))"
-      "          AND ?6 LIKE lens AND ?7 BETWEEN iso_min AND iso_max"
-      "          AND ?8 BETWEEN exposure_min AND exposure_max"
-      "          AND ?9 BETWEEN aperture_min AND aperture_max"
-      "          AND ?10 BETWEEN focal_length_min AND focal_length_max"
-      "          AND (format = 0 OR (format & ?11 != 0 AND ~format & ?12 != 0)))"
-      "        OR (name = ?13))"
-      "   AND operation NOT IN"
-      "        ('ioporder', 'metadata', 'modulegroups', 'export', 'tagging', 'collect', 'basecurve')"
-      " ORDER BY writeprotect DESC, LENGTH(model), LENGTH(maker), LENGTH(lens)",
-      table);
-    // clang-format on
-
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(), query, -1, stmt_ptr, NULL);
-    dt_free(query);
-  }
-
-  sqlite3_stmt *stmt = *stmt_ptr;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 2, image->exif_model, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 3, image->exif_maker, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 4, image->camera_alias, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 5, image->camera_maker, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 6, image->exif_lens, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_DOUBLE(stmt, 7, fmaxf(0.0f, fminf(FLT_MAX, image->exif_iso)));
-  DT_DEBUG_SQLITE3_BIND_DOUBLE(stmt, 8, fmaxf(0.0f, fminf(1000000, image->exif_exposure)));
-  DT_DEBUG_SQLITE3_BIND_DOUBLE(stmt, 9, fmaxf(0.0f, fminf(1000000, image->exif_aperture)));
-  DT_DEBUG_SQLITE3_BIND_DOUBLE(stmt, 10, fmaxf(0.0f, fminf(1000000, image->exif_focal_length)));
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 11, iformat);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 12, excluded);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 13, workflow_preset, -1, SQLITE_TRANSIENT);
-
-  while(sqlite3_step(stmt) == SQLITE_ROW)
-  {
-    const int32_t id = sqlite3_column_int(stmt, 0);
-    const int num = sqlite3_column_int(stmt, 1);
-    const int modversion = sqlite3_column_int(stmt, 2);
-    const char *operation = (const char *)sqlite3_column_text(stmt, 3);
-    const void *module_params = sqlite3_column_blob(stmt, 4);
-    const int enabled = sqlite3_column_int(stmt, 5);
-    const void *blendop_params = sqlite3_column_blob(stmt, 6);
-    const int blendop_version = sqlite3_column_int(stmt, 7);
-    const int multi_priority = sqlite3_column_int(stmt, 8);
-    const char *multi_name = (const char *)sqlite3_column_text(stmt, 9);
-    const char *preset_name = (const char *)sqlite3_column_text(stmt, 10);
-    const int param_length = sqlite3_column_bytes(stmt, 4);
-    const int bl_length = sqlite3_column_bytes(stmt, 6);
-
-    cb(user_data, id, num, modversion, operation, module_params, param_length, enabled,
-       blendop_params, bl_length, blendop_version, multi_priority, multi_name, preset_name);
-  }
-
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-}
-
-gboolean dt_history_db_get_autoapply_ioporder_params(const int32_t imgid, const dt_image_t *image,
-                                                    const int iformat, const int excluded, void **params,
-                                                    int32_t *params_len)
-{
-  if(imgid <= 0 || IS_NULL_PTR(image) || IS_NULL_PTR(params) || !params_len) return FALSE;
-  *params = NULL;
-  *params_len = 0;
-
-  _history_stmt_mutex_ensure();
-  dt_pthread_mutex_lock(&_history_stmt_mutex);
-
-  if(!_history_auto_ioporder_stmt)
-  {
-    // clang-format off
-    DT_DEBUG_SQLITE3_PREPARE_V2(dt_database_get_sqlite3_global(),
-                                "SELECT op_params"
-                                " FROM data.presets"
-                                " WHERE autoapply=1"
-                                "       AND ((?2 LIKE model AND ?3 LIKE maker) OR (?4 LIKE model AND ?5 LIKE maker))"
-                                "       AND ?6 LIKE lens AND ?7 BETWEEN iso_min AND iso_max"
-                                "       AND ?8 BETWEEN exposure_min AND exposure_max"
-                                "       AND ?9 BETWEEN aperture_min AND aperture_max"
-                                "       AND ?10 BETWEEN focal_length_min AND focal_length_max"
-                                "       AND (format = 0 OR (format & ?11 != 0 AND ~format & ?12 != 0))"
-                                "       AND operation = 'ioporder'"
-                                " ORDER BY writeprotect DESC, LENGTH(model), LENGTH(maker), LENGTH(lens)",
-                                -1, &_history_auto_ioporder_stmt, NULL);
-    // clang-format on
-  }
-
-  sqlite3_stmt *stmt = _history_auto_ioporder_stmt;
-  sqlite3_reset(stmt);
-  sqlite3_clear_bindings(stmt);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 1, imgid);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 2, image->exif_model, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 3, image->exif_maker, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 4, image->camera_alias, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 5, image->camera_maker, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_TEXT(stmt, 6, image->exif_lens, -1, SQLITE_TRANSIENT);
-  DT_DEBUG_SQLITE3_BIND_DOUBLE(stmt, 7, fmaxf(0.0f, fminf(FLT_MAX, image->exif_iso)));
-  DT_DEBUG_SQLITE3_BIND_DOUBLE(stmt, 8, fmaxf(0.0f, fminf(1000000, image->exif_exposure)));
-  DT_DEBUG_SQLITE3_BIND_DOUBLE(stmt, 9, fmaxf(0.0f, fminf(1000000, image->exif_aperture)));
-  DT_DEBUG_SQLITE3_BIND_DOUBLE(stmt, 10, fmaxf(0.0f, fminf(1000000, image->exif_focal_length)));
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 11, iformat);
-  DT_DEBUG_SQLITE3_BIND_INT(stmt, 12, excluded);
-
-  gboolean ok = FALSE;
-  if(sqlite3_step(stmt) == SQLITE_ROW)
-  {
-    const void *blob = sqlite3_column_blob(stmt, 0);
-    const int32_t blob_len = sqlite3_column_bytes(stmt, 0);
-    if(blob && blob_len > 0)
-    {
-      *params = g_malloc(blob_len);
-      memcpy(*params, blob, blob_len);
-      *params_len = blob_len;
-      ok = TRUE;
-    }
-  }
-
-  dt_pthread_mutex_unlock(&_history_stmt_mutex);
-  return ok;
-}
 
 #undef DT_IOP_ORDER_INFO
 // clang-format off
