+++ src/darktable.c
@@ -117,6 +117,7 @@
 #include "common/datetime.h"
 #include "common/exif.h"
 #include "common/history.h"
+#include "database/history_repository.h"
 #include "common/pwstorage/pwstorage.h"
 #include "common/selection.h"
 #include "gui/privacy_consent.h"
@@ -1778,7 +1779,7 @@ void dt_cleanup()
   }
 
   dt_colorlabels_cleanup();
-  dt_history_cleanup();
+  dt_history_repository_cleanup();
   dt_dev_history_cleanup();
   dt_metadata_cleanup();
   dt_image_cleanup();
